dev_kit/requirements.txt
setuptools==44.0.0
wheel==0.45.1
click==8.1.7
numpy==1.24.4
matplotlib==3.7.5
pandas==2.0.3
scikit-learn==1.3.2
geopandas==0.13.2
rasterio==1.3.11
pillow==10.4.0
laspy==2.5.4
shapely==2.0.7
fiona==1.9.0
opencv-python==4.10.0.84

